+++ lib/lazy.ts
